Ziga! Look at Me! I have important things! GUI Things!
So, I cleaned out most of the empty render methods in the GUIComponents,
and replaced them with setting the visible boolean to false in
image-less constructors. Its much cleaner that way. There's no point in
having empty methods laying around.

The real thing I need you to look at is Some of the changes I made to
the button.java constructors (they are commented) and the button
intersect method. The intersect method doesn't seem to work because of
camera offsets.

I was trying to get a button set up in the main menu
(Mainmenuscreen.java).

Also, could you comment your methods and classes and explain what
they're for as you go?

<3 from the most annoying person you know, I'm sure.